requirements.txt
langchain
langchain-community
langchain-openai
langchainhub 
azure-ai-documentintelligence
unstructured[all-docs]
langchain-chroma
chromadb
openbb-agents
openbb[all]
yfinance
pyautogen
python-dotenv
xgboost
markdownify
pandas
einops